refactor: split apps/core.py into cohesive modules
core.py had grown to 1701 lines mixing nine concerns behind a flat list
of module-level functions. Split it into four single-purpose modules and
delete core.py (clean break):

- manifest.py  — worker dispatch + git repo cache + manifest discovery/loading
- command.py   — requirement checks + path/param validation + build_command
- jobs.py      — poll loop + job submission/cancel
- jobfiles.py  — work-dir path construction + job file access

Dependencies form a DAG (jobs imports the other three; the rest are
leaves), so there's no circular-import risk. __init__.py re-exports the
same public API, so server.py is unchanged. Repointed user_worker.py
imports, a model.py comment, and test imports/patch targets to the new
modules. Pure reorganization — no behavior change.
